修改打印方法
